TinkerPop 3.0.0M4.
This is a major change to update this package to use TinkerPop 3.
TinkerPop 3 is not backwards compatible to TinkerPop 2. Many changes
are made in this commit to make unit tests pass, but more changes
should be made to clean up the result. For example, the fundamental
concept of a 'pipeline' in TP2 changed to a 'traversal' in TP3,
but this revision still uses the term Pipeline extensively in the
code.