Revert back to toJSON.
While converting to TinkerPop3 I needed to refactor how we convert elements (vertices and edges) to JSON.
In the process I introduced a new method toJsObj, but it should have just been a new implementation
for toJSON. This commit fixes that, and provides sync and async versions.

Also changed bin/gremlin-console, enough to verify that we can at least now execute simple
queries such as 'g.V()' and 'g.E()'.