Add Confirm Traversal API test.
Adds a test that finds a functions of the java GraphTraversal proxy and
compares then against an expected list. This will inform us if the API
changes (not too unlikely, given that the sack API was added in M5), but
more importantly gives us an index of methods that we must provide
wrappers for in order to provide a complete implementation.