Change 'subgraph' API (3.0.0.M8)
Subgraphs are now created for all edges reaching the 'subgraph' step.

Additionally, multiple subgraphs can be routed to the side effect map via the string key.

This changes the gremlin-v3 subgraph step so that it is normalized, and must now be iterated to capture the subgraph.
This is necessary to handle the latest functionality.